requirements.txt
beautifulsoup4
googlesearch-python
